requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.1
ollama==0.4.6
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
